@@ -1,6 +1,6 @@
 plugins {
 	java
-	id("org.springframework.boot") version "2.7.16"
+	id("org.springframework.boot") version "3.2.1"
 	id("io.spring.dependency-management") version "1.0.15.RELEASE"
 }
 
@@ -8,7 +8,7 @@ group = "ru.sber.demo"
 version = "0.0.1-SNAPSHOT"
 
 java {
-	sourceCompatibility = JavaVersion.VERSION_11
+	sourceCompatibility = JavaVersion.VERSION_21
 }
 
 repositories {
